# cmd/user-portal/docs/swagger.yaml
basePath: /
definitions:
  main.GetMeetingTokenResponse:
    properties:
      duration:
        type: integer
      forceEndAt:
        type: string
      meetingId:
        type: string
      participantName:
        type: string
      roomName:
        type: string
      scheduledAt:
        type: string
      token:
        type: string
      url:
        type: string
    type: object
  models.CreateMeetingRequest:
    properties:
      additional_notes:
        description: AdditionalNotes - дополнительные заметки
        example: Подготовить отчеты
        type: string
      department_ids:
        description: DepartmentIDs - идентификаторы отделов
        example:
        - '["550e8400-e29b-41d4-a716-446655440000"]'
        items:
          type: string
        type: array
      duration:
        description: Duration - длительность в минутах
        example: 60
        type: integer
      force_end_at_duration:
        description: ForceEndAtDuration - принудительно завершить встречу после истечения
          времени
        example: true
        type: boolean
      needs_audio_record:
        description: NeedsAudioRecord - требуется ли аудиозапись
        example: true
        type: boolean
      needs_video_record:
        description: NeedsVideoRecord - требуется ли видеозапись
        example: true
        type: boolean
      participant_ids:
        description: ParticipantIDs - идентификаторы участников
        example:
        - '["550e8400-e29b-41d4-a716-446655440000"]'
        items:
          type: string
        type: array
      recurrence:
        allOf:
        - $ref: '#/definitions/models.MeetingRecurrence'
        description: Recurrence - частота повторения
        example: weekly
      scheduled_at:
        description: ScheduledAt - запланированное время начала
        example: "2025-01-15T10:00:00Z"
        type: string
      speaker_id:
        description: SpeakerID - идентификатор докладчика (для презентаций)
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      subject_id:
        description: SubjectID - идентификатор темы
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      title:
        description: Title - название встречи
        example: Еженедельное совещание
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.MeetingType'
        description: Type - тип встречи
        example: conference
    required:
    - duration
    - scheduled_at
    - subject_id
    - title
    - type
    type: object
  models.Department:
    properties:
      created_at:
        description: CreatedAt - время создания отдела
        type: string
      description:
        description: Description - описание отдела
        type: string
      id:
        description: ID - уникальный идентификатор отдела
        type: string
      is_active:
        description: IsActive - активен ли отдел
        type: boolean
      level:
        description: Level - глубина в иерархии (0 для корневого)
        type: integer
      name:
        description: Name - название отдела
        type: string
      parent_id:
        description: ParentID - идентификатор родительского отдела (NULL для корневых
          отделов)
        type: string
      path:
        description: Path - полный путь вида "root/child/grandchild"
        type: string
      updated_at:
        description: UpdatedAt - время последнего обновления отдела
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        description: Code - HTTP код ошибки
        type: integer
      error:
        description: Error - код ошибки
        type: string
      message:
        description: Message - сообщение об ошибке
        type: string
      timestamp:
        description: Timestamp - время возникновения ошибки
        type: string
    type: object
...
  models.MeetingParticipantInfo:
    properties:
      created_at:
        description: CreatedAt - время создания записи
        type: string
      id:
        description: ID - уникальный идентификатор участника
        type: string
      joined_at:
        description: JoinedAt - время присоединения к встрече
        type: string
      left_at:
        description: LeftAt - время выхода из встречи
        type: string
      meeting_id:
        description: MeetingID - идентификатор встречи
        type: string
      role:
        description: Role - роль участника ("speaker", "participant")
        type: string
      status:
        description: Status - статус участника ("invited", "accepted", "declined",
          "attended")
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.UserInfo'
        description: User - информация о пользователе
      user_id:
        description: UserID - идентификатор пользователя
        type: string
    type: object
  models.MeetingRecurrence:
    enum:
    - none
    - daily
    - weekly
    - monthly
    type: string
    x-enum-comments:
      MeetingRecurrenceDaily: Ежедневно
      MeetingRecurrenceMonthly: Ежемесячно
      MeetingRecurrenceNone: Не повторяется
      MeetingRecurrenceWeekly: Еженедельно
    x-enum-descriptions:
    - Не повторяется
    - Ежедневно
    - Еженедельно
    - Ежемесячно
    x-enum-varnames:
    - MeetingRecurrenceNone
    - MeetingRecurrenceDaily
    - MeetingRecurrenceWeekly
    - MeetingRecurrenceMonthly
  models.MeetingStatus:
    enum:
    - scheduled
    - in_progress
    - completed
    - cancelled
    type: string
    x-enum-comments:
      MeetingStatusCancelled: Отменена
      MeetingStatusCompleted: Завершена
      MeetingStatusInProgress: Идет
      MeetingStatusScheduled: Запланирована
    x-enum-descriptions:
    - Запланирована
    - Идет
    - Завершена
    - Отменена
    x-enum-varnames:
    - MeetingStatusScheduled
    - MeetingStatusInProgress
    - MeetingStatusCompleted
    - MeetingStatusCancelled
  models.MeetingSubject:
    properties:
      created_at:
        description: CreatedAt - время создания
        type: string
      department_ids:
        description: DepartmentIDs - отделы, с которыми связана эта тема
        items:
          type: string
        type: array
      description:
        description: Description - описание темы
        type: string
      id:
        description: ID - уникальный идентификатор темы
        type: string
      is_active:
        description: IsActive - активна ли тема
        type: boolean
      name:
        description: Name - название темы
        type: string
      updated_at:
        description: UpdatedAt - время последнего обновления
        type: string
    type: object
...
  models.MeetingType:
    enum:
    - presentation
    - conference
    type: string
    x-enum-comments:
      MeetingTypeConference: Совещание
      MeetingTypePresentation: Доклад
    x-enum-descriptions:
    - Доклад
    - Совещание
    x-enum-varnames:
    - MeetingTypePresentation
    - MeetingTypeConference
  models.MeetingWithDetails:
    properties:
      additional_notes:
        description: AdditionalNotes - дополнительные заметки
        type: string
      created_at:
        description: CreatedAt - время создания
        type: string
      created_by:
        description: CreatedBy - идентификатор пользователя, создавшего встречу
        type: string
      created_by_user:
        allOf:
        - $ref: '#/definitions/models.UserInfo'
        description: CreatedByUser - пользователь, создавший встречу
      departments:
        description: Departments - список отделов
        items:
          $ref: '#/definitions/models.Department'
        type: array
      duration:
        description: Duration - длительность в минутах
        type: integer
      force_end_at_duration:
        description: ForceEndAtDuration - принудительно завершить встречу после истечения
          времени
        type: boolean
      id:
        description: ID - уникальный идентификатор встречи
        type: string
      livekit_room_id:
        description: LiveKitRoomID - ссылка на комнату LiveKit если встреча начата
        type: string
      needs_audio_record:
        description: NeedsAudioRecord - требуется ли аудиозапись
        type: boolean
      needs_video_record:
        description: NeedsVideoRecord - требуется ли видеозапись
        type: boolean
      participants:
        description: Participants - список участников
        items:
          $ref: '#/definitions/models.MeetingParticipantInfo'
        type: array
      recurrence:
        allOf:
        - $ref: '#/definitions/models.MeetingRecurrence'
        description: Recurrence - частота повторения
      scheduled_at:
        description: ScheduledAt - запланированное время начала
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.MeetingStatus'
        description: Status - текущий статус встречи
      subject:
        allOf:
        - $ref: '#/definitions/models.MeetingSubject'
        description: Subject - тема встречи
      subject_id:
        description: SubjectID - идентификатор темы встречи
        type: string
      title:
        description: Title - название встречи
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.MeetingType'
        description: Type - тип встречи
      updated_at:
        description: UpdatedAt - время последнего обновления
        type: string
    type: object
  models.MeetingsResponse:
    properties:
      items:
        description: Items - список встреч
        items:
          $ref: '#/definitions/models.MeetingWithDetails'
        type: array
      offset:
        description: Offset - смещение от начала
        type: integer
      page_size:
        description: PageSize - размер страницы
        type: integer
      total:
        description: Total - общее количество встреч
        type: integer
    type: object
...
  models.UserInfo:
    properties:
      avatar:
        description: Avatar - URL аватара или base64 изображение
        example: https://example.com/avatar.jpg
        type: string
      bio:
        description: Bio - биография пользователя
        example: Software developer
        type: string
      department_id:
        description: DepartmentID - идентификатор отдела, к которому принадлежит пользователь
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      email:
        description: Email - адрес электронной почты пользователя
        example: admin@recontext.online
        type: string
      first_name:
        description: FirstName - имя пользователя
        example: John
        type: string
      id:
        description: ID - уникальный идентификатор пользователя
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      language:
        description: Language - предпочитаемый язык пользователя
        example: en
        type: string
      last_name:
        description: LastName - фамилия пользователя
        example: Doe
        type: string
      permissions:
        allOf:
        - $ref: '#/definitions/models.UserPermissions'
        description: Permissions - специфические разрешения пользователя
      phone:
        description: Phone - номер телефона пользователя
        example: "+1234567890"
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        description: Role - роль пользователя в системе
        example: admin
      username:
        description: Username - имя пользователя для входа
        example: admin
        type: string
    type: object
  models.UserPermissions:
    properties:
      can_approve_recordings:
        description: CanApproveRecordings - разрешение на утверждение записей
        type: boolean
      can_manage_department:
        description: CanManageDepartment - разрешение на управление отделом
        type: boolean
      can_schedule_meetings:
        description: CanScheduleMeetings - разрешение на планирование видеовстреч
        type: boolean
    type: object
  models.UserRole:
    enum:
    - admin
    - user
    - operator
    - service
    type: string
    x-enum-comments:
      RoleService: For service-to-service authentication
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - For service-to-service authentication
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
    - RoleOperator
    - RoleService
...
host: localhost:8081
info:
  contact:
    email: support@recontext.online
    name: API Support
    url: http://recontext.online/support
  description: 'API пользовательского портала Recontext.online: загрузка записей,
    транскрибация, поиск и управление материалами'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://recontext.online/terms/
  title: Recontext.online User Portal API
  version: 0.1.0
paths:
...
  /api/v1/meetings:
    get:
      description: Получить постраничный список встреч, где пользователь является
        участником или докладчиком
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      - description: Фильтр по статусу
        enum:
        - scheduled
        - in_progress
        - completed
        - cancelled
        in: query
        name: status
        type: string
      - description: Фильтр по типу
        enum:
        - presentation
        - conference
        in: query
        name: type
        type: string
      - description: Фильтр по идентификатору темы
        in: query
        name: subject_id
        type: string
      - description: Фильтр по дате начала (формат RFC3339)
        in: query
        name: date_from
        type: string
      - description: Фильтр по дате окончания (формат RFC3339)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MeetingsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список моих встреч
      tags:
      - Meetings
    post:
      consumes:
      - application/json
      description: Создать новую встречу (требуется разрешение can_schedule_meetings
        или роль admin/operator)
      parameters:
      - description: Данные встречи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateMeetingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MeetingWithDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать новую встречу
      tags:
      - Meetings
...
  /api/v1/meetings/{meetingId}/token:
    get:
      consumes:
      - application/json
      description: Получить токен доступа LiveKit для присоединения к комнате встречи
        (можно присоединиться за 10 минут до начала)
      parameters:
      - description: Идентификатор встречи
        in: path
        name: meetingId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GetMeetingTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить токен LiveKit для присоединения к встрече
      tags:
      - Meetings
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"